Rename startup class and fix namespace
